requirements.txt
python-telegram-bot
python-dotenv
deep-translator
speechrecognition
pillow
pytesseract
pydub
gtts